Factor handler functions out of connector module
